Bugfix/9 (#10)
Fix #9 for intermittent failures during batch conversions

- Use a Debian-based container image for entrusted_container
- Disable old mitigations for container image exit codes
- Try to display useful error messages for special error codes
- Update translations
- Update the container security profile for seccomp
- More robust error handling with reasonably clear error message when the PDF result might missing

+++ app/entrusted_client/src/container.rs
@@ -119,18 +119,14 @@ fn exec_crt_command (cmd_desc: String, container_program: common::ContainerProgr
                     return Ok(());
                 } else {
                     if let Some(exit_code) = exit_status.code() {
-                        // Mitigate apparent 139 exit codes that don't happen with Debian bullseye
-                        // Sadly it's a much bigger image than with Alpine so we decide to mitigate the issue
-                        // Apparently a vsyscall=emulate argument needs to be added to /proc/cmdline depending on the kernel version
-                        // Essentially not mitigating the issue would be a problem in a non-controlled environment (i.e. Live CD ISO)
-                        // See https://stackoverflow.com/questions/55508604/docker-is-exited-immediately-when-runs-with-error-code-139
-                        // See https://unix.stackexchange.com/questions/478387/running-a-centos-docker-image-on-arch-linux-exits-with-code-139
-                        if exit_code == 139 {
-                            return Ok(());
-                        }
+                        // Please see https://betterprogramming.pub/understanding-docker-container-exit-codes-5ee79a1d58f6
+                        if exit_code == 139 || exit_code == 137 {
+                            let mut explanation = trans.gettext("Container process terminated abruptly potentially due to memory usage. Are PDF pages too big? Try increasing the container engine memory allocation?");
 
-                        if exit_code == 137 {
-                            let explanation = trans.gettext("Container process terminated abruptly likely due to memory usage. Are PDF pages too big? Trying increasing the container memory.");
+                            if exit_code == 139 {
+                                explanation = trans.gettext("Container process terminated abruptly potentially due to a memory access fault. Please report the issue at: https://github.com/rimerosolutions/entrusted/issues");
+                            }
+
                             let lm = common::LogMessage {
                                 data: format!("{} {}", trans.gettext("Conversion failed!"), explanation),
                                 percent_complete: 100
@@ -277,36 +273,38 @@ pub fn convert(input_path: PathBuf, output_path: PathBuf, convert_options: commo
         dz_tmp.push("entrusted");
         mkdirp(&dz_tmp, trans.clone())?;
         cleanup_dir(&dz_tmp)?;
-        
-        
+
+
         let seccomp_profile_data = include_bytes!("../seccomp-entrusted-profile.json");
         let seccomp_profile_filename = format!("seccomp-entrusted-profile-{}.json", option_env!("CARGO_PKG_VERSION").unwrap_or("Unknown"));
         let seccomp_profile_pathbuf = PathBuf::from(dz_tmp.join(seccomp_profile_filename));
-        convert_args.push("--security-opt".to_string());
-        convert_args.push(format!("seccomp={}", seccomp_profile_pathbuf.display()));
-        
+        // convert_args.push("--security-opt".to_string());
+        // convert_args.push(format!("seccomp={}", seccomp_profile_pathbuf.display()));
+
         if !seccomp_profile_pathbuf.exists() {
             let f_ret = fs::File::create(&seccomp_profile_pathbuf);
-            
+
             match f_ret {
                 Ok(mut f) => {
                     if let Err(ex) = f.write_all(seccomp_profile_data) {
-                        tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;                
+                        tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could not save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;
                         return Err(ex.into());
                     }
-                    
+
                     if let Err(ex) = f.sync_all() {
-                        tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;                
+                        tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could not save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;
                         return Err(ex.into());
                     }
                 },
                 Err(ex) => {
-                    tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;                
+                    tx.send(common::AppEvent::ConversionProgressEvent(printer.print(5, trans.gettext_fmt("Could not save security profile to {0}. {1}.", vec![&seccomp_profile_pathbuf.display().to_string(), &ex.to_string()]))))?;
                     return Err(ex.into());
-                }            
-            }            
+                }
+            }
         }
 
+        // TODO dynamic naming for couple of folders overall
+        // This is needed for parallel conversion and not overwritting files among other things
         let mut dz_tmp_safe:PathBuf = dz_tmp.clone();
         dz_tmp_safe.push("safe");
         mkdirp(&dz_tmp_safe, trans.clone())?;
@@ -349,23 +347,31 @@ pub fn convert(input_path: PathBuf, output_path: PathBuf, convert_options: commo
                 }
             }
 
+            // In the case of a container crash, the output file will not be present...
+            // This should be handled upstream by capturing proper exit codes of the sanitization process
             let mut container_output_file_path = dz_tmp_safe.clone();
             container_output_file_path.push("safe-output-compressed.pdf");
-            let atime = FileTime::now();
 
-            fs::copy(&container_output_file_path, &output_path)?;
-            fs::remove_file(container_output_file_path)?;
+            if !container_output_file_path.exists() {
+                let msg_info = trans.gettext("Potential sanitization process crash detected, the sanitized PDF result was not created.");
+                err_msg.push_str(&msg_info);
+            } else {
+                let atime = FileTime::now();
 
-            let output_file = fs::File::open(&output_path)?;
+                fs::copy(&container_output_file_path, &output_path)?;
+                fs::remove_file(container_output_file_path)?;
 
-            // This seems to fail on Microsoft Windows with permission denied errors
-            let _ = filetime::set_file_handle_times(&output_file, Some(atime), Some(atime));
+                let output_file = fs::File::open(&output_path)?;
 
-            if let Err(ex) = cleanup_dir(&dz_tmp_safe) {
-                tx.send(common::AppEvent::ConversionProgressEvent(printer.print(100, trans.gettext_fmt("Failed to cleanup temporary folder: {0}. {1}.", vec![&dz_tmp.clone().display().to_string(), &ex.to_string()]))))?;
+                // This seems to fail on Microsoft Windows with permission denied errors
+                let _ = filetime::set_file_handle_times(&output_file, Some(atime), Some(atime));
+
+                if let Err(ex) = cleanup_dir(&dz_tmp_safe) {
+                    tx.send(common::AppEvent::ConversionProgressEvent(printer.print(100, trans.gettext_fmt("Failed to cleanup temporary folder: {0}. {1}.", vec![&dz_tmp.clone().display().to_string(), &ex.to_string()]))))?;
+                }
+
+                success = true;
             }
-
-            success = true;
         } else {
             err_msg = trans.gettext("Conversion failed!");
         }
